+++ src/components/Accordion/index.ts
@@ -1,2 +1,3 @@
+"use client";
 export { default, Accordion } from "./Accordion";
 export type { AccordionProps, AccordionItemProps } from "./Accordion";