+++ src/components/Dropdown/index.ts
@@ -1,2 +1,3 @@
+"use client";
 export { default, Dropdown } from "./Dropdown";
 export type { DropdownProps, DropdownItemProps } from "./Dropdown";