+++ src/components/Tooltip/index.ts
@@ -1,2 +1,3 @@
+"use client";
 export { default, Tooltip } from "./Tooltip";
 export type { TooltipProps, TooltipPlacement } from "./Tooltip";